调整： 1、底层调整：映射关系调整，底层框架分离，IStartup启动器实现。 2、ORM调整：新增EF、Dapper支持。原ORM使用Linq合并修改功能，支持异步处理，命名区分为Lts（更好的使用数据库连接池【支持配置最大连接数】）。

+++ Tests/CodeArts.Db.EntityFramework.Tests/Domain/Entities/FeiUsers.cs
@@ -52,7 +52,7 @@ namespace UnitTest.Domain.Entities
         /// <summary>
         /// 角色组。
         /// </summary>
-        public int Mallagid { get; set; }
+        public short Mallagid { get; set; }
 
         /// <summary>
         /// 盐。
